requirements.txt
pyspark
elasticsearch
kibana
kafka-python
numpy
